requirements.txt
flask
tensorflow
streamlit
pandas
numpy
pillow
requests
